snap: Revert yaru commit causing incompatibilities with <= GTK 4.15.2
The problem appears to be the use of CSS relative colors in 86549dd, which
requires GTK >= 4.15.2 (released on 28th June). It looks like the older alpha()
and shade() functions continue to work in current GTK releases, so reverting
that commit might be enough.

See: https://github.com/ubuntu/yaru/issues/4148

Co-Authored-By: James Henstridge <james@jamesh.id.au>